+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -6,8 +6,8 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
-* Added `AddLink(SpanContext, SpanAttributes?)` to `TelemetrySpan` to support linking spans and associating optional attributes
-  for advanced trace relationships.
+* Added `AddLink(SpanContext, SpanAttributes?)` to `TelemetrySpan` to support
+  linking spans and associating optional attributes for advanced trace relationships.
   ([#6305](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6305))
 
 ## 1.12.0